requirements.txt
arrow==0.15.7
atomicwrites==1.4.0
attrs==19.3.0
azure-cognitiveservices-vision-computervision==0.6.0
azure-cognitiveservices-vision-face==0.4.1
azure-common==1.1.25
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudsync==1.5.1
cloudsync-onedrive==1.0.1
colorama==0.4.3
cycler==0.10.0
dataclasses==0.6
docutils==0.16
idna==2.9
importlib-metadata==1.6.1
isodate==0.6.0
kiwisolver==1.2.0
lockfile==0.12.2
matplotlib==3.2.2
more-itertools==8.4.0
msgpack==1.0.0
msrest==0.6.16
numpy==1.19.0
oauthlib==3.1.0
onedrivesdk-fork==1.1.14
packaging==20.4
pathtools==0.1.2
Pillow==7.1.2
pluggy==0.13.1
py==1.8.2
pyparsing==2.4.7
pystrict==1.1
pytest==5.4.3
python-daemon==2.2.4
python-dateutil==2.8.1
pywin32==228
PyYAML==5.3.1
quickxorhash==1.0.4
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.25.9
watchdog==0.10.2
wcwidth==0.2.4
xxhash==1.4.4
zipp==3.1.0
